requirements.txt
PyYAML==6.0
pytest==7.4.0
pytest-mock==3.10.0
jsonschema==4.21.4
